locals working

+++ include/parametrics.hpp
@@ -3,32 +3,32 @@
 
 class Coord {
     public:
-        float x, y;
+        double x, y;
 
-        Coord(float x, float y)
+        Coord(double x, double y)
             : x(x),
               y(y) {}
 
         Coord operator+(const Coord& other) const;
         Coord operator-(const Coord& other) const;
-        float operator*(const Coord& other) const;
+        double operator*(const Coord& other) const;
         Coord operator*(const float& scalar) const;
         Coord operator/(const float& scalar) const;
         Coord &operator+=(const Coord &other);
         Coord lerp(Coord other, float t) const;
-        float distance(const Coord& other) const;
-        float angle(const Coord& other) const;
-        Coord rotate(float angle) const;
+        double distance(const Coord& other) const;
+        double angle(const Coord& other) const;
+        Coord rotate(double angle) const;
 };
 
 class Pose : public Coord {
     public:
-    float theta;
+    double theta;
 
-    Pose(float x, float y, float theta) : Coord(x, y), theta(theta) {}
+    Pose(double x, double y, double theta) : Coord(x, y), theta(theta) {}
     
     // Constructor to convert pxl::Coord to Pose
     Pose(const Coord &other) : Coord(other), theta() {}
    
-    float getCurvature(Pose& pose, Pose& other);
+    double getCurvature(Pose& pose, Pose& other);
 };
